clean up store
.
