+++ src/renderer/utils/hooks/typedUseSelector.ts
@@ -2,6 +2,6 @@ import {
   TypedUseSelectorHook,
   useSelector as useReduxSelector,
 } from 'react-redux'
-import { GlobalState } from '@/renderer/store/rootReducer'
+import { GlobalState } from '@/renderer/store/store'
 
 export const useSelector: TypedUseSelectorHook<GlobalState> = useReduxSelector
