+++ src/renderer/utils/ros/rosClient.ts
@@ -1,5 +1,6 @@
 import RosClient from './roslib-ts-client/RosClient'
 import { rosService } from '@/renderer/state/ros'
+import { store } from '@/renderer/store/store'
 
 const { IP, port } = rosService.state.context
 export const rosClient = new RosClient(IP, port)
